fixes for clang

+++ wowViewerLib/src/engine/algorithms/mathHelper.cpp
@@ -257,12 +257,12 @@ mathfu::vec3 MathHelper::getIntersection( mathfu::vec3 &tail1, mathfu::vec3 &hea
 
     mathfu::vec3 npoints[2];//Nearest points' coordinates
 
-    npoints[0] = tail1 + dir1 * sc;
-    npoints[1] = tail2 + dir2 * tc;
+    npoints[0] = tail1 + dir1 * (float)sc;
+    npoints[1] = tail2 + dir2 * (float)tc;
 
     if((npoints[1] - npoints[0]).Length() < 0.1)
     {
-        return (npoints[0] + npoints[1]) / 2;
+        return (npoints[0] + npoints[1]) / 2.0f;
     }
     else return mathfu::vec3(0,0,0);
 }
@@ -424,7 +424,7 @@ void MathHelper::sortVec3ArrayAgainstPlane(std::vector<mathfu::vec3> &thisPortal
     for (int j = 0; j < thisPortalVertices.size(); j++) {
         center += thisPortalVertices[j];
     }
-    center *= 1.0 / thisPortalVertices.size();
+    center *= 1.0f / (float)thisPortalVertices.size();
 
     std::sort(thisPortalVertices.begin(), thisPortalVertices.end(),
         [&](mathfu::vec3 &a, mathfu::vec3 &b) -> bool {
